+++ src/operadores/q33inversoSoma.java
@@ -5,11 +5,11 @@ import java.util.Scanner;
 public class q33inversoSoma {
 
     public static void main(String[] args) {
-
+        //Entrada:
         Scanner input = new Scanner(System.in);
 
         System.out.println("Numero de 3 digitos: ");
-        int num = input.nextInt();
+        double num = input.nextInt();
 
         //Processamento:
         double c = Math.floor(num / 100);
